% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_utils.r
\name{deduplicate_xlsx_files_in_dir}
\alias{deduplicate_xlsx_files_in_dir}
\title{Identifies duplicate xlsx files, copies only unique files
Version: 03
Date: 2025-09-08 UTC}
\usage{
deduplicate_xlsx_files_in_dir(path_in, path_out, copy = TRUE)
}
\arguments{
\item{path_in}{Path to directory with xlsx files}

\item{path_out}{Path to directory for copied unique files}

\item{copy}{Logical; if FALSE, do not copy files, only report duplicates}
}
\value{
Invisibly, a list with elements \code{duplicates} and \code{copied}.
}
\description{
Identifies duplicate xlsx files, copies only unique files
Version: 03
Date: 2025-09-08 UTC
}
\examples{
deduplicate_xlsx_files_in_dir(
  path_in  = "~/Downloads/tmp",
  path_out = "~/tmp/",
  copy     = TRUE
)
}
\author{
Toni Wöhrl
}
